refactor: group helpers in one file, split router, clean types
